Reorganize CLI scripts into single directory
