+++ disdrodb/cli/disdrodb_run_l0c.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 # -----------------------------------------------------------------------------.
 # Copyright (c) 2021-2023 DISDRODB developers
 #
@@ -14,16 +15,13 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # -----------------------------------------------------------------------------.
-"""Script to run the DISDRODB L0 processing."""
+"""Script to run the DISDRODB L0C processing."""
 import sys
 from typing import Optional
 
 import click
 
-from disdrodb.l0.routines import (
-    click_l0_archive_options,
-)
-from disdrodb.utils.scripts import (
+from disdrodb.utils.cli import (
     click_base_dir_option,
     click_processing_options,
     click_stations_options,
@@ -36,20 +34,13 @@ sys.tracebacklimit = 0  # avoid full traceback error if occur
 
 @click.command()
 @click_stations_options
-@click_l0_archive_options
 @click_processing_options
 @click_base_dir_option
-def disdrodb_run_l0(
+def disdrodb_run_l0c(
     # L0 disdrodb stations options
     data_sources: Optional[str] = None,
     campaign_names: Optional[str] = None,
     station_names: Optional[str] = None,
-    # L0 archive options
-    l0a_processing: bool = True,
-    l0b_processing: bool = True,
-    l0b_concat: bool = True,
-    remove_l0a: bool = False,
-    remove_l0b: bool = False,
     # Processing options
     force: bool = False,
     verbose: bool = True,
@@ -58,7 +49,7 @@ def disdrodb_run_l0(
     base_dir: Optional[str] = None,
 ):
     """
-    Run the L0 processing of DISDRODB stations.
+    Run the L0C processing of DISDRODB stations.
 
     This function allows to launch the processing of many DISDRODB stations with a single command.
     From the list of all available DISDRODB stations, it runs the processing of the
@@ -66,9 +57,6 @@ def disdrodb_run_l0(
 
     Parameters
     ----------
-    base_dir : str
-        Base directory of DISDRODB
-        Format: <...>/DISDRODB
     data_sources : str
         Name of data source(s) to process.
         The name(s) must be UPPER CASE.
@@ -81,24 +69,6 @@ def disdrodb_run_l0(
     station_names : str
         Station names.
         To specify multiple stations, write i.e.: --station_names 'station1 station2'
-    l0a_processing : bool
-        Whether to launch processing to generate L0A Apache Parquet file(s) from raw data.
-        The default is True.
-    l0b_processing : bool
-        Whether to launch processing to generate L0B netCDF4 file(s) from L0A data.
-        The default is True.
-    l0b_concat : bool
-        Whether to concatenate all raw files into a single L0B netCDF file.
-        If l0b_concat=True, all raw files will be saved into a single L0B netCDF file.
-        If l0b_concat=False, each raw file will be converted into the corresponding L0B netCDF file.
-        The default is False.
-    remove_l0a : bool
-        Whether to keep the L0A files after having generated the L0B netCDF products.
-        The default is False.
-    remove_l0b : bool
-         Whether to remove the L0B files after having concatenated all L0B netCDF files.
-         It takes places only if l0b_concat = True
-        The default is False.
     force : bool
         If True, overwrite existing data into destination directories.
         If False, raise an error if there are already data into destination directories.
@@ -110,20 +80,19 @@ def disdrodb_run_l0(
         If True, the files are processed simultaneously in multiple processes.
         Each process will use a single thread to avoid issues with the HDF/netCDF library.
         By default, the number of process is defined with os.cpu_count().
-        However, you can customize it by typing i.e. DASK_NUM_WORKERS=4 disdrodb_run_l0
+        However, you can customize it by typing: DASK_NUM_WORKERS=4 disdrodb_run_l0b
         If False, the files are processed sequentially in a single process.
         If False, multi-threading is automatically exploited to speed up I/0 tasks.
     debugging_mode : bool
         If True, it reduces the amount of data to process.
-        For L0A, it processes just the first 3 raw data files.
-        For L0B, it processes just the first 100 rows of 3 L0A files.
+        It processes just the first 100 rows of 3 L0A files for each station.
         The default is False.
     base_dir : str
         Base directory of DISDRODB
         Format: <...>/DISDRODB
         If not specified, uses path specified in the DISDRODB active configuration.
     """
-    from disdrodb.l0.routines import run_disdrodb_l0
+    from disdrodb.l0.routines import run_disdrodb_l0c
 
     # Parse data_sources, campaign_names and station arguments
     base_dir = parse_base_dir(base_dir)
@@ -132,17 +101,11 @@ def disdrodb_run_l0(
     station_names = parse_arg_to_list(station_names)
 
     # Run processing
-    run_disdrodb_l0(
+    run_disdrodb_l0c(
         base_dir=base_dir,
         data_sources=data_sources,
         campaign_names=campaign_names,
         station_names=station_names,
-        # L0 archive options
-        l0a_processing=l0a_processing,
-        l0b_processing=l0b_processing,
-        l0b_concat=l0b_concat,
-        remove_l0a=remove_l0a,
-        remove_l0b=remove_l0b,
         # Processing options
         force=force,
         verbose=verbose,
