+++ disdrodb/cli/disdrodb_run_l0c_station.py
@@ -20,7 +20,7 @@ from typing import Optional
 
 import click
 
-from disdrodb.utils.scripts import (
+from disdrodb.utils.cli import (
     click_base_dir_option,
     click_processing_options,
     click_station_arguments,
