+++ disdrodb/utils/cli.py
@@ -16,7 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # -----------------------------------------------------------------------------.
-"""DISDRODB scripts utility."""
+"""DISDRODB command-line-interface scripts utilities."""
 
 import click
