+++ pyproject.toml
@@ -70,29 +70,29 @@ include = ["disdrodb*"]
 
 [project.scripts]
 # Initialization
-disdrodb_initialize_station="disdrodb.api.scripts.disdrodb_initialize_station:disdrodb_initialize_station"
+disdrodb_initialize_station="disdrodb.cli.disdrodb_initialize_station:disdrodb_initialize_station"
 # Metadata archive
-disdrodb_check_metadata_archive="disdrodb.metadata.scripts.disdrodb_check_metadata_archive:disdrodb_check_metadata_archive"
+disdrodb_check_metadata_archive="disdrodb.cli.disdrodb_check_metadata_archive:disdrodb_check_metadata_archive"
 # Data transfer
-disdrodb_download_archive="disdrodb.data_transfer.scripts.disdrodb_download_archive:disdrodb_download_archive"
-disdrodb_download_station="disdrodb.data_transfer.scripts.disdrodb_download_station:disdrodb_download_station"
-disdrodb_upload_archive="disdrodb.data_transfer.scripts.disdrodb_upload_archive:disdrodb_upload_archive"
-disdrodb_upload_station="disdrodb.data_transfer.scripts.disdrodb_upload_station:disdrodb_upload_station"
+disdrodb_download_archive="disdrodb.cli.disdrodb_download_archive:disdrodb_download_archive"
+disdrodb_download_station="disdrodb.cli.disdrodb_download_station:disdrodb_download_station"
+disdrodb_upload_archive="disdrodb.cli.disdrodb_upload_archive:disdrodb_upload_archive"
+disdrodb_upload_station="disdrodb.cli.disdrodb_upload_station:disdrodb_upload_station"
 # L0A
-disdrodb_run_l0a_station="disdrodb.l0.scripts.disdrodb_run_l0a_station:disdrodb_run_l0a_station"
-disdrodb_run_l0a="disdrodb.l0.scripts.disdrodb_run_l0a:disdrodb_run_l0a"
+disdrodb_run_l0a_station="disdrodb.cli.disdrodb_run_l0a_station:disdrodb_run_l0a_station"
+disdrodb_run_l0a="disdrodb.cli.disdrodb_run_l0a:disdrodb_run_l0a"
 # L0B
-disdrodb_run_l0b_station="disdrodb.l0.scripts.disdrodb_run_l0b_station:disdrodb_run_l0b_station"
-disdrodb_run_l0b="disdrodb.l0.scripts.disdrodb_run_l0b:disdrodb_run_l0b"
+disdrodb_run_l0b_station="disdrodb.cli.disdrodb_run_l0b_station:disdrodb_run_l0b_station"
+disdrodb_run_l0b="disdrodb.cli.disdrodb_run_l0b:disdrodb_run_l0b"
 # L0C
-disdrodb_run_l0c_station="disdrodb.l0.scripts.disdrodb_run_l0c_station:disdrodb_run_l0c_station"
-disdrodb_run_l0c="disdrodb.l0.scripts.disdrodb_run_l0c:disdrodb_run_l0c"
+disdrodb_run_l0c_station="disdrodb.cli.disdrodb_run_l0c_station:disdrodb_run_l0c_station"
+disdrodb_run_l0c="disdrodb.cli.disdrodb_run_l0c:disdrodb_run_l0c"
 # L0
-disdrodb_run_l0_station="disdrodb.l0.scripts.disdrodb_run_l0_station:disdrodb_run_l0_station"
-disdrodb_run_l0="disdrodb.l0.scripts.disdrodb_run_l0:disdrodb_run_l0"
+disdrodb_run_l0_station="disdrodb.cli.disdrodb_run_l0_station:disdrodb_run_l0_station"
+disdrodb_run_l0="disdrodb.cli.disdrodb_run_l0:disdrodb_run_l0"
 # L0B concatenation
-disdrodb_run_l0b_concat_station="disdrodb.l0.scripts.disdrodb_run_l0b_concat_station:disdrodb_run_l0b_concat_station"
-disdrodb_run_l0b_concat="disdrodb.l0.scripts.disdrodb_run_l0b_concat:disdrodb_run_l0b_concat"
+disdrodb_run_l0b_concat_station="disdrodb.cli.disdrodb_run_l0b_concat_station:disdrodb_run_l0b_concat_station"
+disdrodb_run_l0b_concat="disdrodb.cli.disdrodb_run_l0b_concat:disdrodb_run_l0b_concat"
